pyproject.toml
[tool.poetry]
name = "cts-diff"
version = "0.1.0"
description = ""
authors = ["Declan McNamara <declan@umich.edu>"]
readme = "README.md"
packages = [{include = "cts_diff"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
jax = "^0.5.0"
einops = "^0.8.0"
hydra-core = "^1.3.2"
equinox = "^0.11.11"
numpyro = "^0.16.1"
lightning = "^2.5.0.post0"
optax = "^0.2.4"
beartype = "^0.19.0"
matplotlib = "^3.10.0"
cloudpickle = "^3.1.1"
torchvision = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
